asyncio_chat/requirements.txt
aiosqlite==0.18.0
pytest-asyncio==0.20.3
aiofiles==22.1.0